Shuffle some tests to more appropriate places
