+++ test/join-tests.js
@@ -23,6 +23,7 @@ var util = require('util');
 var test2 = require('./test-util').test2;
 var dsl = require('./ringpop-assert');
 var prepareCluster = require('./test-util').prepareCluster;
+var prepareWithStatus = require('./test-util').prepareWithStatus;
 var getClusterSizes = require('./it-tests').getClusterSizes;
 
 function joinFakeCluster(n) {
@@ -52,4 +53,35 @@ test2('join ringpop with fake node', getClusterSizes(), 20000,
     ];})
 );
 
+function joinFrom(n, status, incNoDelta, deltaAlive, nSuspect, nFaulty) {
+    test2('join from ' + status + ' with incNoDelta ' + incNoDelta, n, 20000, 
+        prepareWithStatus(0, status, function(t, tc, n) {
+            return [
+                dsl.disableNode(t, tc, 0),
+                dsl.enableNode(t, tc, 0, tc.fakeNodes[0].incarnationNumber+incNoDelta),
+                dsl.sendJoin(t, tc, 0),
+                dsl.waitForJoinResponse(t, tc, 0),
+                // We expect the node to not accept the join but not change it's own membership
+                // A node is expected to disseminate its own existence
+                function(list, cb) {
+                    tc.fakeNodes[0].incarnationNumber -= incNoDelta;
+                    cb(list);
+                },
+                dsl.assertStats(t, tc, n + deltaAlive, nSuspect, nFaulty, {0: {status: status}}),
+            ];
+        })
+    );
+}
+
+joinFrom(getClusterSizes(), 'alive', -1, 1, 0, 0);
+joinFrom(getClusterSizes(), 'alive',  0, 1, 0, 0);
+joinFrom(getClusterSizes(), 'alive',  1, 1, 0, 0);
+
+joinFrom(clusterSizes, 'suspect', -1, 0, 1, 0);
+joinFrom(clusterSizes, 'suspect',  0, 0, 1, 0);
+joinFrom(clusterSizes, 'suspect',  1, 0, 1, 0);
+
+joinFrom(clusterSizes, 'faulty', -1, 0, 0, 1);
+joinFrom(clusterSizes, 'faulty',  0, 0, 0, 1);
+joinFrom(clusterSizes, 'faulty',  1, 0, 0, 1);
