+++ test/ping-tests.js
@@ -22,6 +22,7 @@ var events = require('./events');
 var test2 = require('./test-util').test2;
 var dsl = require('./ringpop-assert');
 var prepareCluster = require('./test-util').prepareCluster;
+var prepareWithStatus = require('./test-util').prepareWithStatus;
 var getClusterSizes = require('./it-tests').getClusterSizes;
 var _ = require('lodash');
 
@@ -40,3 +41,59 @@ test2('ping ringpop from fake-nodes', getClusterSizes(), 20000,
 	    ];
 	})
 );
+
+// piggyback {alive, suspect, faulty} status of fake-node
+// who is {alive, suspect, faulty} with {lower, equal, higher}
+// incarnation number than the fake-node (27 combinations)
+function changeStatus(ns, initial, newState, finalState, incNoDelta, deltaAlive, nSuspect, nFaulty) {
+    var ix = 1;
+    test2('change status from ' + initial + ', to ' + newState + 
+        ' with incNoDelta ' + incNoDelta + ' via piggybacking', 
+        ns, 20000, prepareWithStatus(ix, initial, function(t, tc, n) {
+            expectedMembers = {}
+            expectedMembers[ix] = {status: finalState};
+            return [
+                dsl.sendPing(t, tc, 0, 
+                    {sourceIx: 0, subjectIx: ix, status: newState, subjectIncNoDelta: incNoDelta}),
+                dsl.waitForPingResponse(t, tc, 0),
+                dsl.assertStats(t, tc, n + deltaAlive, nSuspect, nFaulty, expectedMembers),
+            ];
+        })
+    );
+}
+
+changeStatus(getClusterSizes(), 'alive',  'alive', 'alive', -1, 1, 0, 0);
+changeStatus(getClusterSizes(), 'alive',  'alive', 'alive',  0, 1, 0, 0);
+changeStatus(getClusterSizes(), 'alive',  'alive', 'alive',  1, 1, 0, 0);
+
+changeStatus(getClusterSizes(), 'alive',  'suspect', 'alive',  -1, 1, 0, 0);
+changeStatus(getClusterSizes(), 'alive',  'suspect', 'suspect', 0, 0, 1, 0);
+changeStatus(getClusterSizes(), 'alive',  'suspect', 'suspect', 1, 0, 1, 0);
+
+changeStatus(getClusterSizes(), 'alive',  'faulty', 'alive', -1, 1, 0, 0);
+changeStatus(getClusterSizes(), 'alive',  'faulty', 'faulty', 0, 0, 0, 1);
+changeStatus(getClusterSizes(), 'alive',  'faulty', 'faulty', 1, 0, 0, 1);
+
+changeStatus(getClusterSizes(), 'suspect', 'alive', 'suspect', -1, 0, 1, 0);
+changeStatus(getClusterSizes(), 'suspect', 'alive', 'suspect',  0, 0, 1, 0);
+changeStatus(getClusterSizes(), 'suspect', 'alive', 'alive',   1, 1, 0, 0);
+
+changeStatus(getClusterSizes(), 'suspect', 'suspect', 'suspect', -1, 0, 1, 0);
+changeStatus(getClusterSizes(), 'suspect', 'suspect', 'suspect', 0,  0, 1, 0);
+changeStatus(getClusterSizes(), 'suspect', 'suspect', 'suspect', 1,  0, 1, 0);
+
+changeStatus(getClusterSizes(), 'suspect', 'faulty', 'suspect', -1, 0, 1, 0);
+changeStatus(getClusterSizes(), 'suspect', 'faulty', 'faulty',  0,  0, 0, 1);
+changeStatus(getClusterSizes(), 'suspect', 'faulty', 'faulty',  1,  0, 0, 1);
+
+changeStatus(getClusterSizes(), 'faulty',  'alive', 'faulty', -1, 0, 0, 1);
+changeStatus(getClusterSizes(), 'faulty',  'alive', 'faulty', 0,  0, 0, 1);
+changeStatus(getClusterSizes(), 'faulty',  'alive', 'alive',  1,  1, 0, 0);
+
+changeStatus(getClusterSizes(), 'faulty',  'suspect', 'faulty', -1, 0, 0, 1);
+changeStatus(getClusterSizes(), 'faulty',  'suspect', 'faulty',  0, 0, 0, 1);
+changeStatus(getClusterSizes(), 'faulty',  'suspect', 'suspect', 1, 0, 1, 0);
+
+changeStatus(getClusterSizes(), 'faulty',  'faulty', 'faulty', -1, 0, 0, 1);
+changeStatus(getClusterSizes(), 'faulty',  'faulty', 'faulty',  0, 0, 0, 1);
+changeStatus(getClusterSizes(), 'faulty',  'faulty', 'faulty',  1, 0, 0, 1);
